feat(dto): 메뉴 그룹 리팩터링
