+++ src/test/java/kitchenpos/menu/domain/menugroup/MenuGroupAcceptanceTest.java
@@ -9,13 +9,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import static kitchenpos.menu.domain.fixture.MenuGroupDomainFixture.메뉴_그룹_생성_요청;
+import static kitchenpos.menu.domain.fixture.MenuGroupDomainFixture.메뉴_그룹_조회_요청;
 import static kitchenpos.utils.AcceptanceFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayName("인수 테스트 - 메뉴그룹 관리")
 class MenuGroupAcceptanceTest extends AcceptanceTest {
@@ -25,24 +26,32 @@ class MenuGroupAcceptanceTest extends AcceptanceTest {
     private MenuGroupResponse 일인_세트_등록됨;
     private MenuGroupResponse 패밀리_세트_등록됨;
 
-
-    public static ExtractableResponse<Response> 메뉴_그룹_생성_요청(MenuGroupRequest menuGroupRequest) {
-        return post("/new/api/menu-groups", menuGroupRequest);
-    }
-
-    public static ExtractableResponse<Response> 메뉴_그룹_조회_요청() {
-        return get("/new/api/menu-groups");
-    }
-
     private void 메뉴_그룹_생성됨(ExtractableResponse<Response> actual) {
         MenuGroupResponse response = actual.as(MenuGroupResponse.class);
         assertThat(response.getName()).isEqualTo(일인_세트.getName());
     }
 
     private void 메뉴_그룹_조회됨(ExtractableResponse<Response> actual, MenuGroupResponse... expected) {
-        List<MenuGroupResponse> response = actual.jsonPath().getList(".",MenuGroupResponse.class);
-        assertThat(response).contains(expected);
+        List<String> expectedNames = Arrays.stream(expected).map(MenuGroupResponse::getName).collect(Collectors.toList());
 
+        List<MenuGroupResponse> response = actual.jsonPath().getList(".", MenuGroupResponse.class);
+        List<String> responseNames = response.stream().map(MenuGroupResponse::getName).collect(Collectors.toList());
+
+        assertThat(responseNames).containsAll(expectedNames);
+    }
+
+    @Test
+    @DisplayName("메뉴그룹 조회")
+    public void 메뉴_그룹_조회() {
+        // given
+        일인_세트_등록됨 = 메뉴_그룹_생성_요청(MenuGroupRequest.from("1인 세트")).as(MenuGroupResponse.class);
+        패밀리_세트_등록됨 = 메뉴_그룹_생성_요청(MenuGroupRequest.from("패밀리 세트")).as(MenuGroupResponse.class);
+
+        // when
+        ExtractableResponse<Response> actual = 메뉴_그룹_조회_요청();
+
+        응답_OK(actual);
+        메뉴_그룹_조회됨(actual, 일인_세트_등록됨, 패밀리_세트_등록됨);
     }
 
     @Nested
@@ -75,21 +84,6 @@ class MenuGroupAcceptanceTest extends AcceptanceTest {
             응답_BAD_REQUEST(actual);
         }
     }
-
-    @Test
-    @DisplayName("메뉴그룹 조회")
-    public void 메뉴_그룹_조회() {
-        // given
-        일인_세트_등록됨 = 메뉴_그룹_생성_요청(MenuGroupRequest.from("1인 세트")).as(MenuGroupResponse.class);
-        패밀리_세트_등록됨 =  메뉴_그룹_생성_요청(MenuGroupRequest.from("패밀리 세트")).as(MenuGroupResponse.class);
-
-        // when
-        ExtractableResponse<Response> actual = 메뉴_그룹_조회_요청();
-
-        응답_OK(actual);
-        메뉴_그룹_조회됨(actual, 일인_세트_등록됨, 패밀리_세트_등록됨);
-    }
-
 
 
 }
